Added more documentation and arranged code
