+++ One-Electron-Property/Restrained-Electrostatic-Potential/README.md
@@ -1,0 +1,29 @@
+Restrained-Electrostatic-Potential
+==================================
+
+The following codes are available:
+- `espfit.py`: The restrained electrostatic potential (RESP) fitting procedure.
+- `resp_driver.py`: Driver of the code.
+- `example.py`: Example for two-stage charge fitting for one conformer.
+- `example2.py`: Example for two-stage charge fitting for two conformers.
+
+Helper programs:
+- `resp_helper.py`: A collection of helper classes and function for RESP. 
+
+RESP charges result from fitting the classical electrostatic potential (ESP)
+generated by atom-centered point charges to the quantum ESP computed outside
+the van der Waals surface of the molecule. The fitted charges are restrained
+by a hyperbolic term, which requires an iterative fitting procedure to compute
+the charges.
+
+The charges (**q**) are computed by solving the following equation:
+**A** **q** = **B**.
+
+The left-hand side contains the information about the classical ESP while the
+right-hand side contains information about the quantum ESP.
+A hyperbolic restraint term that depends on the charges is added to the diagonal
+elements of matrix **A**. The charges and the diagonal elements of **A** are
+changed iteratively until the charges converge.
+
+Reference:
+Bayly:93:10269-10280(https://pubs.acs.org/doi/abs/10.1021/j100142a004) C. I. Bayly *et. al.* *JPC* **97**, 10269-10280 (1993)
